third_party/blink/web_tests/platform/generic/fast/speechsynthesis/speech-synthesis-mock-no-crash-expected.txt

This tests that the mock speech synthesizer is well behaved when attempting to pause/resume without first speaking.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS speechSynthesis.pause() is undefined.
PASS speechSynthesis.resume() is undefined.
PASS successfullyParsed is true

TEST COMPLETE
